update for emp. vehicle sync

+++ parking-repository/src/main/java/com/spm/repository/VehicleRepository.java
@@ -17,7 +17,7 @@ public interface VehicleRepository extends JpaRepository<VehicleEntity, Long>{
 	
 	List<VehicleEntity> findAllByProjectId(long projectId);
 	
-	VehicleEntity findByVehicleIdAndProjectId(long type,  long projectId);
+	VehicleEntity findByVehicleIdAndProjectId(int type,  long projectId);
 	
 	VehicleEntity findById(long id);
